gateway/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic
python-multipart==0.0.6
alembic
sqlalchemy
passlib
jose
psycopg2-binary

# Google Cloud & Firebase
google-cloud-firestore==2.13.1
google-cloud-speech==2.23.0
google-cloud-texttospeech==2.16.4
google-cloud-vision==3.4.5
google-ai-generativelanguage==0.4.0
firebase-admin==6.2.0

# HTTP & API
httpx==0.25.2
requests==2.31.0

# Data processing
pandas
numpy==1.26.2

# Environment & Configuration
python-dotenv==1.0.0
pyyaml==6.0.1

# Development & Testing
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
flake8==6.1.0

# Other utilities
python-dateutil==2.8.2
pytz==2023.3
